requirements.txt
asgiref==3.9.1
Django==5.2.5
django-debug-toolbar==6.0.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.1
whitenoise==6.9.0
